+++ test/unit/mysql_rep_admin/unit_test_run.sh
@@ -22,3 +22,4 @@ test/unit/mysql_rep_admin/process_time_lag.py
 test/unit/mysql_rep_admin/rpt_mst_log.py
 test/unit/mysql_rep_admin/rpt_slv_log.py
 test/unit/mysql_rep_admin/run_program.py
+test/unit/mysql_rep_admin/transpose_dict.py